requirements.txt
Flask == 2.1.0
pyyaml
pandas == 1.5.2
Flask-Log-Request-ID
connexion == 2.14.0
prometheus-client==0.11.0
prometheus-flask-exporter==0.20.3
